workaround for 1st boot hangs
long story short - at the time we configured a new interface "systemctl
restart systemd-networkd" was invoked by our guest plugin.
That implied that the internal interface to vagrant would be reloaded,
but since we had set a (new) hostname in the meantime that would induce
a new IP to be issued (100% certain in VMware lately) - which would mess
big time with the early provisioning phase - instead of keeping the boot
one. this hack afaict fixes it for good.

Signed-off-by: António Meireles <antonio.meireles@reformi.st>