Drop containers-basic bundle from the default list.
This patch is dropping containers-basic (mainly docker) from the default
image. While docker is used massively across the world, it will not be
used on all scenarios of this image. Also docker muddles with iptable
rules which is not needed for all applications and use-cases. It can
still be installed by running `swupd bundle-add containers-basic` if an
user requires it.

Signed-off-by: Ganesh Maharaj Mahalingam <ganesh.mahalingam@intel.com>